+++ src/idunno.Bluesky/Notifications/BlueskyAgent.cs
@@ -5,7 +5,6 @@ using System.Diagnostics.CodeAnalysis;
 
 using idunno.AtProto;
 using idunno.AtProto.Repo;
-using idunno.Bluesky.Actor;
 using idunno.Bluesky.Notifications;
 
 namespace idunno.Bluesky;
@@ -35,59 +34,6 @@ public partial class BlueskyAgent
             loggerFactory: LoggerFactory,
             cancellationToken: cancellationToken).ConfigureAwait(false);
     }
-
-    /// <summary>
-    /// Gets the notifications for the requesting account. Requires authentication.
-    /// </summary>
-    /// <param name="subscribedLabelers">A optional list of labeler <see cref="Did"/>s to accept labels from.</param>
-    /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
-    /// <returns>The task object representing the asynchronous operation.</returns>
-    /// <exception cref="AuthenticationRequiredException">Thrown when the current agent is not authenticated.</exception>
-    public async Task<AtProtoHttpResult<NotificationCollection>> ListNotifications(IEnumerable<Did>? subscribedLabelers = null, CancellationToken cancellationToken = default)
-    {
-        if (!IsAuthenticated)
-        {
-            throw new AuthenticationRequiredException();
-        }
-
-        return await ListNotifications(null, null, null, subscribedLabelers, cancellationToken).ConfigureAwait(false);
-    }
-
-    /// <summary>
-    /// Gets the notifications for the requesting account. Requires authentication.
-    /// </summary>
-    /// <param name="limit">The maximum number of notifications to return. If specified this should be &gt;=1 and &lt;= 100.</param>
-    /// <param name="cursor">An optional cursor. See https://atproto.com/specs/xrpc#cursors-and-pagination.</param>
-    /// <param name="seenAt">The date and time notifications were last checked.</param>
-    /// <param name="subscribedLabelers">A optional list of labeler <see cref="Did"/>s to accept labels from.</param>
-    /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
-    /// <returns>The task object representing the asynchronous operation.</returns>
-    /// <exception cref="AuthenticationRequiredException">Thrown when the current agent is not authenticated.</exception>
-    public async Task<AtProtoHttpResult<NotificationCollection>> ListNotifications(
-        int? limit = null,
-        string? cursor = null,
-        DateTimeOffset? seenAt = null,
-        IEnumerable<Did>? subscribedLabelers = null,
-        CancellationToken cancellationToken = default)
-    {
-        if (!IsAuthenticated)
-        {
-            throw new AuthenticationRequiredException();
-        }
-
-        return await BlueskyServer.ListNotifications(
-            limit,
-            cursor,
-            seenAt,
-            service: Service,
-            accessCredentials: Credentials,
-            httpClient: HttpClient,
-            onCredentialsUpdated: InternalOnCredentialsUpdatedCallBack,
-            loggerFactory: LoggerFactory,
-            subscribedLabelers: subscribedLabelers,
-            cancellationToken: cancellationToken).ConfigureAwait(false);
-    }
-
 
     /// <summary>
     /// Updates the date and time notifications were last seen for the current user. Requires authentication.
@@ -144,42 +90,6 @@ public partial class BlueskyAgent
 
         return
             await GetBlueskyRecord<Declaration>($"at://{did}/{CollectionNsid.NotificationDeclaration}/self", cancellationToken: cancellationToken).ConfigureAwait(false);
-    }
-
-    /// <summary>
-    /// Gets the activity subscriptions for the requesting account.
-    /// </summary>
-    /// <param name="limit">The maximum number of activity subscriptions to return. If specified this should be greater than 1 and less than or equal to 100.</param>
-    /// <param name="cursor">An optional cursor. See https://atproto.com/specs/xrpc#cursors-and-pagination.</param>
-    /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
-    /// <returns>The task object representing the asynchronous operation.</returns>
-    /// <exception cref="ArgumentOutOfRangeException">Thrown when <paramref name="limit"/>&lt;1 or <paramref name="limit"/>&gt;100.</exception>
-    /// <exception cref="AuthenticationRequiredException">Thrown when the current session is not authenticated.</exception>
-    public async Task<AtProtoHttpResult<PagedViewReadOnlyCollection<ProfileView>>> ListActivitySubscriptions(
-        int? limit = null,
-        string? cursor = null,
-        CancellationToken cancellationToken = default)
-    {
-        if (limit is not null)
-        {
-            ArgumentOutOfRangeException.ThrowIfLessThan((int)limit, 1);
-            ArgumentOutOfRangeException.ThrowIfGreaterThan((int)limit, 100);
-        }
-
-        if (!IsAuthenticated)
-        {
-            throw new AuthenticationRequiredException();
-        }
-
-        return await BlueskyServer.ListActivitySubscriptions(
-            limit: limit,
-            cursor: cursor,
-            service: Service,
-            accessCredentials: Credentials,
-            httpClient: HttpClient,
-            onCredentialsUpdated: InternalOnCredentialsUpdatedCallBack,
-            loggerFactory: LoggerFactory,
-            cancellationToken: cancellationToken).ConfigureAwait(false);
     }
 
     /// <summary>